Logic rewrite
Moved most of the logic to server so that the server is in control of what views the clients render. This refactor fixes a lot of edge cases like people dropping out of the queue, active player dropping out midst game etc. The logic and state of the game is still inside the game for easier pluggability.

PS: The server code is horrible and wet atm. I'll clean it up tomorrow.