Vibrate when it's your turn (#84)
* Extract controller state observable

* Vibrate phone when it's your turn

Note: the vibration API is implemented a bit differently on different
browsers. On iOS, vibrate doesn't work at all because the
navigator.vibrate function is undefined. On Android Chrome, the user
must first tap somewhere on the page before vibrate works. On Android
Firefox, the browser will show a confirm box to the user asking whether
to allow the page to use vibration.