Remove unused command listener from withSocket
Originally, the commands from the controller would be delegated to the
display which computed the game logic. Nowadays the logic is handled on
the server, meaning that the clients do not need to listen to the
command events at all - they only listen for "state" events.

Removed unnecessary socket event listener from withSocket, stuff
related to the old architecture from storybook, and the
SocketIODisplays::sendActions method from the server.